Fix GH Windows Companion build.
