+++ radio/src/translations/string_list.h
@@ -107,9 +107,10 @@ STR(MAIN_MENU_MODEL_NOTES)
 STR(MAIN_MENU_MODEL_SETTINGS)
 STR(MAIN_MENU_RADIO_SETTINGS)
 STR(MAIN_MENU_SCREEN_SETTINGS)
-STR(MAIN_MENU_SETTINGS)
 STR(MAIN_MENU_STATISTICS)
 STR(MAIN_MENU_THEMES)
+STR(MAIN_MODEL_SETTINGS)
+STR(MAIN_RADIO_SETTINGS)
 STR(MANAGE_MODELS)
 STR(MEM_USED_EXTRA)
 STR(MEM_USED_SCRIPT)
@@ -191,7 +192,6 @@ STR(QM_CURVES)
 STR(QM_CUSTOM_LUA)
 STR(QM_DEBUG)
 STR(QM_FLIGHT_MODES)
-STR(QM_GEN_SETTINGS)
 STR(QM_GLOB_FUNC)
 STR(QM_GLOBAL_VARS)
 STR(QM_HARDWARE)
@@ -200,8 +200,10 @@ STR(QM_LOGICAL_SW)
 STR(QM_LS_MON)
 STR(QM_MANAGE_MODELS)
 STR(QM_MIXES)
+STR(QM_MODEL_SETTINGS)
 STR(QM_MODEL_SETUP)
 STR(QM_OUTPUTS)
+STR(QM_RADIO_SETTINGS)
 STR(QM_RADIO_SETUP)
 STR(QM_RESET)
 STR(QM_SCREEN_1)
@@ -223,6 +225,19 @@ STR(QM_TOOLS)
 STR(QM_TOP_BAR)
 STR(QM_TRAINER)
 STR(QM_UI_SETUP)
+STR(QUICK_MENU_FAVORITES)
+STR(CURRENT_SCREEN)
+STR(MAIN_MENU_APPS)
+STR(MAIN_VIEW_1)
+STR(MAIN_VIEW_10)
+STR(MAIN_VIEW_2)
+STR(MAIN_VIEW_3)
+STR(MAIN_VIEW_4)
+STR(MAIN_VIEW_5)
+STR(MAIN_VIEW_6)
+STR(MAIN_VIEW_7)
+STR(MAIN_VIEW_8)
+STR(MAIN_VIEW_9)
 // Main Menu
 
 #else // !COLORLCD
@@ -586,6 +601,7 @@ STR(JACK_MODE)
 STR(JITTER_FILTER)
 STR(KEYS_BTN)
 STR(KEYSTUCK)
+STR(KEY_SHORTCUTS)
 STR(LCD)
 STR(LENGTH)
 STR(LIMITS_HEADERS_DIRECTION)
@@ -594,19 +610,9 @@ STR(LIMITS_HEADERS_SUBTRIM)
 STR(LIMITS_HEADERS_SUBTRIMMODE)
 STR(LOADINGMODEL)
 STR(LOGS)
+STR(LONG_PRESS)
 STR(LOWALARM)
 STR(LUA_SCRIPTS_LABEL)
-STR(MAIN_MENU_APPS)
-STR(MAIN_VIEW_1)
-STR(MAIN_VIEW_10)
-STR(MAIN_VIEW_2)
-STR(MAIN_VIEW_3)
-STR(MAIN_VIEW_4)
-STR(MAIN_VIEW_5)
-STR(MAIN_VIEW_6)
-STR(MAIN_VIEW_7)
-STR(MAIN_VIEW_8)
-STR(MAIN_VIEW_9)
 STR(MAX)
 STR(MAXBAUDRATE)
 STR(MEMORYWARNING)
@@ -766,6 +772,7 @@ STR(OFF)
 STR(OFFSET)
 STR(OK)
 STR(ONLYPOSITIVE)
+STR(OPEN_QUICK_MENU)
 STR(OPTIONS)
 STR(OTA_UPDATE_ERROR)
 STR(OTA_UPDATE)
@@ -911,6 +918,7 @@ STR(SF_VARIO)
 STR(SF_VOLUME)
 STR(SF)
 STR(SHARE)
+STR(SHORT_PRESS)
 STR(SHOW_INSTANCE_ID)
 STR(SIDE)
 STR(SIGNAL_OUTPUT)
